Move `BuildOptions` to the `options` submodule
Just code motion.
